+++ agent/handlers/v1/response_test.go
@@ -109,7 +109,7 @@ func TestTaskResponse(t *testing.T) {
 	}
 
 	containerNameToDockerContainer := map[string]*apicontainer.DockerContainer{
-		taskARN: &apicontainer.DockerContainer{
+		taskARN: {
 			DockerID:   containerID,
 			DockerName: containerName,
 			Container:  container,