+++ agent/wsclient/client.go
@@ -120,6 +120,9 @@ type ClientServerImpl struct {
 	// message with said message. It will be called before a RequestHandler is
 	// called. It must take a single interface{} argument.
 	AnyRequestHandler RequestHandler
+	// MakeRequestHook is an optional callback that, if set, is called on every
+	// generated request with the raw request body.
+	MakeRequestHook MakeRequestHookFunc
 	// URL is the full url to the backend, including path, querystring, and so on.
 	URL string
 	// RWTimeout is the duration used for setting read and write deadlines
@@ -131,6 +134,11 @@ type ClientServerImpl struct {
 	ServiceError
 	TypeDecoder
 }
+
+// MakeRequestHookFunc is a function that is invoked on every generated request
+// with the raw request body.  MakeRequestHookFunc must return either the body
+// to send or an error.
+type MakeRequestHookFunc func([]byte) ([]byte, error)
 
 // Connect opens a connection to the backend and upgrades it to a websocket. Calls to
 // 'MakeRequest' can be made after calling this, but responses will not be
@@ -326,6 +334,13 @@ func (cs *ClientServerImpl) MakeRequest(input interface{}) error {
 	send, err := cs.CreateRequestMessage(input)
 	if err != nil {
 		return err
+	}
+
+	if cs.MakeRequestHook != nil {
+		send, err = cs.MakeRequestHook(send)
+		if err != nil {
+			return err
+		}
 	}
 
 	// Over the wire we send something like