+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
+## 1.20.3
+* Enhancement - Deprecate support for serial docker image pull [#1569](https://github.com/aws/amazon-ecs-agent/pull/1569)
+* Enhancement - Update the `amazon-ecs-cni-plugins` to `2018.08.0`
 
-## 1.20.2-dev
+## 1.20.2
 * Enhancement - Added ECS config field `ECS_SHARED_VOLUME_MATCH_FULL_CONFIG` to
 make the volume labels and driver options comparison configurable for shared volume [#1519](https://github.com/aws/amazon-ecs-agent/pull/1519)
 * Enhancement - Added Volumes metadata as part of v1 and v2 metadata endpoints [#1531](https://github.com/aws/amazon-ecs-agent/pull/1531)